pi/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pillar-controller"
version = "1.1.0"
description = "LED Pillar Controller — Raspberry Pi application"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.104",
  "uvicorn[standard]>=0.24",
  "pyserial>=3.5",
  "numpy>=1.26",
  "Pillow>=10.0",
  "pyyaml>=6.0",
  "websockets>=12.0",
  "python-multipart>=0.0.6",
  "astral>=3.2",
]

[project.optional-dependencies]
audio = ["sounddevice>=0.4"]
video = ["av>=12.0"]
dev = ["pytest>=7.0", "pytest-asyncio>=0.23", "httpx>=0.25"]

[tool.setuptools.packages.find]
include = ["app*"]

[project.scripts]
pillar = "app.main:main"
